refactor: move models to the domain layer
