+++ lib/domain/models/day.dart
