+++ lib/domain/models/enum/frequency.dart
