+++ lib/domain/models/enum/settings.dart
