+++ lib/domain/models/history_entry.dart
