+++ lib/pages/settings/settings_update_sleep_schedule_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:hidroly/data/model/enum/frequency.dart';
-import 'package:hidroly/data/model/enum/settings.dart';
+import 'package:hidroly/domain/models/enum/frequency.dart';
+import 'package:hidroly/domain/models/enum/settings.dart';
 import 'package:hidroly/l10n/app_localizations.dart';
 import 'package:hidroly/provider/settings_provider.dart';
 import 'package:hidroly/services/notification_service.dart';
