+++ lib/provider/settings_provider.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:hidroly/data/model/enum/frequency.dart';
-import 'package:hidroly/data/model/enum/settings.dart';
+import 'package:hidroly/domain/models/enum/frequency.dart';
+import 'package:hidroly/domain/models/enum/settings.dart';
 import 'package:hidroly/utils/app_date_utils.dart';
 import 'package:shared_preferences/shared_preferences.dart';
