+++ lib/services/notification_service.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
-import 'package:hidroly/data/model/enum/settings.dart';
+import 'package:hidroly/domain/models/enum/settings.dart';
 import 'package:hidroly/l10n/app_localizations.dart';
 import 'package:hidroly/utils/app_date_utils.dart';
 import 'package:workmanager/workmanager.dart';
