+++ lib/widgets/home/water_action_buttons.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:hidroly/data/model/history_entry.dart';
-import 'package:hidroly/data/model/water_button.dart';
+import 'package:hidroly/domain/models/history_entry.dart';
+import 'package:hidroly/domain/models/water_button.dart';
 import 'package:hidroly/l10n/app_localizations.dart';
 import 'package:hidroly/provider/app_state_provider.dart';
 import 'package:hidroly/provider/custom_cups_provider.dart';
